backend/src/advanced_features.py
"""Placeholders for advanced features like ControlNet and inpainting."""

from typing import Optional
from PIL import Image


class AdvancedImageGenerator:
    def __init__(self, base_generator):
        self.base_generator = base_generator

    def generate_with_controlnet(
        self,
        prompt: str,
        control_image: Image.Image,
        controlnet_type: str = "canny",
        controlnet_conditioning_scale: float = 1.0,
    ):
        """Stub for ControlNet generation."""
        # In a full implementation, apply controlnet to the base generator
        return self.base_generator.generate_image(prompt)

    def generate_img2img(
        self,
        prompt: str,
        init_image: Image.Image,
        strength: float = 0.8,
    ):
        """Stub for image-to-image generation."""
        return self.base_generator.generate_image(prompt)

    def generate_inpainting(
        self,
        prompt: str,
        image: Image.Image,
        mask_image: Image.Image,
    ):
        """Stub for inpainting generation."""
        return self.base_generator.generate_image(prompt)

