backend/requirements.txt
fastapi
uvicorn
pydantic
Pillow
diffusers
transformers
accelerate
torch
